conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.1
aiohappyeyeballs=2.4.3=pyhd8ed1ab_0
aiohttp=3.10.9=py312h024a12e_0
aiosignal=1.3.1=pyhd8ed1ab_0
annotated-types=0.7.0=pyhd8ed1ab_0
attrs=24.2.0=pyh71513ae_0
beautifulsoup4=4.12.3=pyha770c72_0
brotli-python=1.0.9=py312h313beb8_8
bzip2=1.0.8=h80987f9_6
c-ares=1.34.1=hd74edd7_0
ca-certificates=2024.8.30=hf0a4a13_0
cachetools=5.5.0=pyhd8ed1ab_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py312h0fad829_0
charset-normalizer=3.4.0=pyhd8ed1ab_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=43.0.1=py312h3ddc590_0
curl=8.10.1=h13a7ad3_0
expat=2.6.3=hf9b8971_0
freetype=2.12.1=hadb7bae_2
frozenlist=1.4.1=py312h024a12e_1
gdbm=1.18=h8fe7016_4
gettext=0.21.0=h13f89a0_1
git=2.47.0=pl5321hc14f901_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.43=pyhd8ed1ab_0
google-ai-generativelanguage=0.6.10=pyhd8ed1ab_0
google-api-core=2.20.0=pyhd8ed1ab_0
google-api-python-client=2.148.0=pyhff2d567_0
google-auth=2.35.0=pyhff2d567_0
google-auth-httplib2=0.2.0=pyhd8ed1ab_0
google-generativeai=0.8.2=pyhd8ed1ab_0
googleapis-common-protos=1.65.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httplib2=0.22.0=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=72.1=he12128b_0
idna=3.10=pyhd8ed1ab_0
iniconfig=1.1.1=pyhd3eb1b0_0
joblib=1.4.2=pypi_0
krb5=1.21.3=h237132a_0
lcms2=2.16=ha0e7c42_0
lerc=4.0.0=h9a09cb3_0
libabseil=20240116.2=cxx17_h313beb8_0
libblas=3.9.0=17_osxarm64_openblas
libcblas=3.9.0=17_osxarm64_openblas
libcurl=8.10.1=h13a7ad3_0
libcxx=19.1.1=ha82da77_0
libdeflate=1.22=hd74edd7_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libexpat=2.6.3=hf9b8971_0
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
libiconv=1.17=h0d3ecfb_2
libintl=0.22.5=h8414b35_3
libjpeg-turbo=3.0.0=hb547adb_1
liblapack=3.9.0=17_osxarm64_openblas
libnghttp2=1.58.0=ha4dd798_1
libopenblas=0.3.23=openmp_hc731615_0
libpng=1.6.44=hc14010f_0
libprotobuf=5.27.4=h02e2e85_0
libsqlite=3.46.0=hfb93653_0
libssh2=1.11.0=h7a5bd25_0
libtiff=4.7.0=hfce79cd_1
libwebp-base=1.4.0=h93a5062_0
libxcb=1.17.0=hdb1d25a_0
libxml2=2.10.4=h2aff0a6_0
libzlib=1.3.1=h8359307_2
llvm-openmp=14.0.6=hc6e5704_0
multidict=6.1.0=py312h88e9dfa_0
ncurses=6.5=h7bae524_1
nltk=3.8.1=pypi_0
numpy=2.1.2=py312h801f5e3_0
openjpeg=2.5.2=h9f1df11_0
openssl=3.3.2=h8359307_0
packaging=24.1=py312hca03da5_0
pcre2=10.44=h297a79d_2
perl=5.34.0=h1a28f6b_2
pillow=10.4.0=py312h8609ca0_1
pip=24.0=py312hca03da5_0
pluggy=1.0.0=py312hca03da5_1
proto-plus=1.24.0=py312hca03da5_0
protobuf=5.27.4=py312hde4cb15_1
pthread-stubs=0.4=hd74edd7_1002
pyasn1=0.6.1=pyhd8ed1ab_1
pyasn1-modules=0.4.1=pyhd8ed1ab_0
pycparser=2.22=pyhd8ed1ab_0
pydantic=2.9.2=pyhd8ed1ab_0
pydantic-core=2.23.4=py312he431725_0
pyopenssl=24.2.1=pyhd8ed1ab_2
pyparsing=3.1.4=pyhd8ed1ab_0
pysocks=1.7.1=pyha2e5f31_6
pytest=7.4.4=py312hca03da5_0
python=3.12.2=hdf0ec26_0_cpython
python_abi=3.12=5_cp312
pyu2f=0.1.5=pyhd8ed1ab_0
pyyaml=6.0.2=py312h024a12e_1
readline=8.2=h1a28f6b_0
regex=2024.7.24=pypi_0
requests=2.32.3=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
setuptools=69.5.1=py312hca03da5_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.0=pyhd8ed1ab_0
soupsieve=2.5=pyhd8ed1ab_1
sqlite=3.46.0=h5838104_0
tk=8.6.13=h5083fa2_1
tqdm=4.66.4=pypi_0
truecase=0.0.14=pypi_0
types-beautifulsoup4=4.12.0.20240907=pyhd8ed1ab_0
types-html5lib=1.1.11.20240806=pyhd8ed1ab_0
types-pillow=10.2.0.20240822=pyhd8ed1ab_0
types-pyyaml=6.0.12.20240917=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
uritemplate=4.1.1=pyhd8ed1ab_0
urllib3=2.2.3=pyhd8ed1ab_0
wheel=0.43.0=py312hca03da5_0
xorg-libxau=1.0.11=hd74edd7_1
xorg-libxdmcp=1.1.5=hd74edd7_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h3422bc3_2
yarl=1.13.1=py312h024a12e_0
zlib=1.3.1=h8359307_2
zstandard=0.23.0=py312h15fbf35_1
zstd=1.5.6=hb46c0d2_0
